requirements.txt
certifi==2019.6.16
chardet==3.0.4
dnspython==1.16.0
idna==2.8
PyJWT==1.7.1
requests==2.22.0
Salesforce-FuelSDK==1.3.0
suds-jurko==0.6
urllib3==1.25.3
